r2783: Making the debugger work with input files larger than 2G, and making it load them faster with buffered I/O.

+++ util/debug/BadInputException.java
@@ -1,26 +1,26 @@
 public class BadInputException extends Exception
 {
-	public final int offset;
+	public final long offset;
 	
-	public BadInputException(int offset)
+	public BadInputException(long offset)
 	{
 		super();
 		this.offset = offset;
 	}
 	
-	public BadInputException(String message, int offset)
+	public BadInputException(String message, long offset)
 	{
 		super(message);
 		this.offset = offset;
 	}
 	
-	public BadInputException(Throwable cause, int offset)
+	public BadInputException(Throwable cause, long offset)
 	{
 		super(cause);
 		this.offset = offset;
 	}
 	
-	public BadInputException(String message, Throwable cause, int offset)
+	public BadInputException(String message, Throwable cause, long offset)
 	{
 		super(message, cause);
 		this.offset = offset;
