+++ util/debug/MissingParameterException.java
@@ -3,7 +3,7 @@ public class MissingParameterException extends BadInputException
 	public final String name;
 	public final int size;
 	
-	public MissingParameterException(String name, int size, int offset)
+	public MissingParameterException(String name, int size, long offset)
 	{
 		super("Missing parameter: \"" + name + "\" of size " + size, offset);
 		this.name = name;
