+++ util/debug/UnexpectedOpcodeException.java
@@ -2,7 +2,7 @@ public class UnexpectedOpcodeException extends BadInputException
 {
 	public final short opcodeNumber;
 	
-	public UnexpectedOpcodeException(short opcodeNumber, int offset)
+	public UnexpectedOpcodeException(short opcodeNumber, long offset)
 	{
 		super("Unexpected opcode: " + Opcode.hex(opcodeNumber), offset);
 		this.opcodeNumber = opcodeNumber;
